dengue_prediction/features/contrib/user_456/feature_1234.py
import sklearn.preprocessing

from fhub_transformers.nans import LagImputer

input = 'ndvi_ne'
transformer = [
    LagImputer(groupby_kwargs={'level': 'city'}),
    sklearn.preprocessing.Imputer(),
    sklearn.preprocessing.StandardScaler(),
]
