requirements.txt
click
Sphinx
awscli
python-dotenv>=0.5.1
category_encoders
fhub_core>=0.1.0
fhub_transformers>=0.2.2
funcy
numpy
pandas
scikit_learn
sklearn_pandas
git+https://github.com/micahjsmith/BTB.git@48db0da94a28492220ca666de80e193c50d756f6#egg=btb-0.2.0
